requirements.txt
yfinance>=0.2.40
openai>=2.0.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
numpy>=1.24.0
